requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-tavily
langgraph
python-dotenv
requests
beautifulsoup4
markdownify
pytest
pytest-mock
